refactor CardPools BoulderElectionOAsim extends BoulderElectionOA, handles both CLCA and OA with cvrs
