+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/oneaudit/BallotPool.kt
@@ -1,0 +1,58 @@
+package org.cryptobiotic.rlauxe.oneaudit
+
+import io.github.oshai.kotlinlogging.KotlinLogging
+import org.cryptobiotic.rlauxe.util.VotesAndUndervotes
+import org.cryptobiotic.rlauxe.util.margin2mean
+import org.cryptobiotic.rlauxe.util.mean2margin
+
+
+private val logger = KotlinLogging.logger("BallotPool")
+
+// this is really CardPoolForContest
+data class BallotPool(
+    val name: String,
+    val poolId: Int,
+    val contestId :Int,
+    val ncards: Int,          // ncards for this contest in this pool; TODO hasStyles = false?
+    val votes: Map<Int, Int>, // candid -> nvotes, for plurality. umm do we really need ?
+) {
+    // TODO does this really agree with the average assorter?
+    // this could go from -1 to 1. TODO shouldnt that be -u to u ??
+    fun calcReportedMargin(winner: Int, loser: Int): Double {
+        if (ncards == 0) return 0.0
+        val winnerVote = votes[winner] ?: 0
+        val loserVote = votes[loser] ?: 0
+        return (winnerVote - loserVote) / ncards.toDouble()
+    }
+
+    fun votesAndUndervotes(voteForN: Int, ncandidates: Int): Map<Int, Int> {
+        val poolVotes = votes.values.sum()
+        val poolUndervotes = ncards * voteForN - poolVotes
+        return (votes.map { Pair(it.key, it.value)} + Pair(ncandidates, poolUndervotes)).toMap()
+    }
+
+    fun votesAndUndervotes(voteForN: Int): VotesAndUndervotes {
+        val poolUndervotes = ncards * voteForN - votes.values.sum()
+        return VotesAndUndervotes(votes, poolUndervotes, voteForN)
+    }
+
+    fun reportedAverage(winner: Int, loser: Int): Double {
+        val winnerVotes = votes[winner] ?: 0
+        val loserVotes = votes[loser] ?: 0
+        val reportedMargin = (winnerVotes - loserVotes) / ncards.toDouble() // TODO dont know Nc
+        return margin2mean(reportedMargin)
+    }
+}
+
+
+// for calculating average from running total, see addOAClcaAssorters
+class AssortAvg() {
+    var ncards = 0
+    var totalAssort = 0.0
+    fun avg() : Double = if (ncards == 0) 0.0 else totalAssort / ncards
+    fun margin() : Double = mean2margin(avg())
+
+    override fun toString(): String {
+        return "AssortAvg(ncards=$ncards, totalAssort=$totalAssort avg=${avg()})"
+    }
+}
